Replace NSegment::isCH with NSegment::whatsit which tells what this segment is part of: heap, stack, break area. So it's a generalisation of sorts. This is handy occasionally and also improves the segment listing.
The change affects the mergeability of segments. Previously, the
SkAnonC segment that makes up the mapped part of the extensible client stack
could be merged with an adjacent SkAnonC segment. This is no longer the case.
And likewise for the SkAnonC segment that is the client break area.
I'm not unduly concerned. Even before this change it was possible for
two neighbouring SkAnonC segments to exist, namely, if one was part of the
client heap (isCH == 1) and the other was not (isCH == 0).
So it was never (well.... since r4789) true that segments were "maximally
merged".


git-svn-id: svn://svn.valgrind.org/valgrind/branches/ASPACEM_TWEAKS@15223 a5019735-40e9-0310-863c-91ae7b9d1cf9